Seperated Feeder and Indexer in code
