+++ src/main/java/frc/robot/subsystems/feeder/FeederBehavior.java
@@ -1,0 +1,23 @@
+package frc.robot.subsystems.feeder;
+
+import frc.robot.util.AllEvents;
+import frc.robot.util.SubsystemBehavior;
+
+public class FeederBehavior extends SubsystemBehavior {
+  private final FeederSubsystem feeder;
+
+  public FeederBehavior(FeederSubsystem feeder) {
+    this.feeder = feeder;
+  }
+
+  @Override
+  public void configure(AllEvents events) {
+    events.goals().isIdleTrigger().whileTrue(feeder.idleCommand());
+
+    events
+        .goals()
+        .isShootingTrigger()
+        .whileTrue(feeder.indexingCommand())
+        .whileFalse(feeder.idleCommand());
+  }
+}
