+++ src/main/java/frc/robot/subsystems/feeder/FeederEvents.java
@@ -1,0 +1,9 @@
+package frc.robot.subsystems.feeder;
+
+import edu.wpi.first.wpilibj2.command.button.Trigger;
+
+public interface FeederEvents {
+  public Trigger isIdleTrigger();
+
+  public Trigger isIndexingTrigger();
+}
