+++ src/main/java/frc/robot/subsystems/indexer/IndexerState.java
@@ -7,28 +7,18 @@ import frc.robot.util.LoggedTunableNumber;
 import java.util.function.DoubleSupplier;
 
 public enum IndexerState {
-  TESTING(() -> 0.0, () -> 0.0),
-  IDLE(() -> 0.0, () -> 0.0),
-  FEEDING(
-      new LoggedTunableNumber("Indexer/setIndexerPoint", 160),
-      new LoggedTunableNumber("Indexer/setFeederPoint", 4000)),
-  REVERSING(
-      new LoggedTunableNumber("Indexer/setIndexerPointReverse", -50),
-      new LoggedTunableNumber("Indexer/setFeederPointReverse", -2000));
+  TESTING(() -> 0.0),
+  IDLE(() -> 0.0),
+  FEEDING(new LoggedTunableNumber("Indexer/setIndexerPoint", 160)),
+  REVERSING(new LoggedTunableNumber("Indexer/setIndexerPointReverse", -50));
 
   private DoubleSupplier m_indexerVelocity;
-  private DoubleSupplier m_feederVelocity;
 
-  private IndexerState(DoubleSupplier indexerVelocity, DoubleSupplier feederVelocity) {
+  private IndexerState(DoubleSupplier indexerVelocity) {
     m_indexerVelocity = indexerVelocity;
-    m_feederVelocity = feederVelocity;
   }
 
   public AngularVelocity indexerVelocity() {
     return RPM.of(m_indexerVelocity.getAsDouble());
-  }
-
-  public AngularVelocity feederVelocity() {
-    return RPM.of(m_feederVelocity.getAsDouble());
   }
 }
